refactor: 폴더 구조 변경, 라우트 연결 수정
- Pages에 ItemDetail 폴더를 생성하고 관련 파일 이동
- Pages 아래의 각 폴더의 이름 첫글자 대문자로 수정
- 이름 변경에 따라 라우트 연결 수정
