+++ src/pages/FreeBoard/FreeBoard.jsx
@@ -1,0 +1,7 @@
+export default function FreeBoard() {
+  return (
+    <>
+      <h1>FreeBoard Page</h1>
+    </>
+  );
+}
