--- src/pages/board/Board.jsx
@@ -1,7 +1,0 @@
-export default function Board() {
-  return (
-    <>
-      <h1>Board Page</h1>
-    </>
-  );
-}
